restructure a bit and try adding tests
