+++ exercises/CMakeLists.txt
@@ -4,12 +4,13 @@ project(VAMR)
 # set_property(TARGET tgt PROPERTY CXX_STANDARD 17)
 set (CMAKE_CXX_STANDARD 17)
 
+# enables configuration generation for `ctest` command
+enable_testing()
 
 ################ 3rd Party Libraries ################
 find_package (Eigen3 REQUIRED NO_MODULE)
 find_package(OpenCV REQUIRED)
 include_directories(${OpenCV_INCLUDE_DIRS})
-
 
 
 ################## Learning 3rd party libraries ##################
@@ -40,18 +41,28 @@ target_link_libraries(utils Eigen3::Eigen ${OpenCV_LIBS})
 
 
 ########################### Exercises ###########################
-# exercise 01
+###### exercise 01
 add_executable(exercise01 exercise01.cpp)
-target_link_libraries (exercise01 utils Eigen3::Eigen ${OpenCV_LIBS})
+target_link_libraries (exercise01 utils)
 
-# exercise 02
+###### exercise 02
 add_executable(exercise02 exercise02.cpp)
-target_link_libraries (exercise02 utils Eigen3::Eigen ${OpenCV_LIBS})
+target_link_libraries (exercise02 utils)
 
-# exercise 03
+###### exercise 03
 add_executable(exercise03 exercise03.cpp)
 target_link_libraries (exercise03 utils Eigen3::Eigen ${OpenCV_LIBS})
 
-# exercise 03
+###### exercise 03
+add_library(sift sift.cpp sift.hpp)
+get_filename_component(SIFT_HEADER_DIR "./" ABSOLUTE)
+target_include_directories(sift PUBLIC ${SIFT_HEADER_DIR})
+target_link_libraries (sift utils)
+
 add_executable(exercise04 exercise04.cpp)
-target_link_libraries (exercise04 utils Eigen3::Eigen ${OpenCV_LIBS})
+target_link_libraries (exercise04 sift)
+
+
+
+############################ Unit Tests ###########################
+add_subdirectory(tests)
