+++ exercises/tests/test_sift.cpp
@@ -1,0 +1,7 @@
+#include "sift.hpp"
+#include <gtest/gtest.h>
+
+TEST(SiftTest, TestTest)
+{
+    EXPECT_TRUE(false);
+}
